Update toast - Display only once toast - Remove toast.go file and move the toast widget to component
