+++ ui/page.go
@@ -5,7 +5,6 @@ import (
 	"image"
 	"image/color"
 	"net/http"
-	"time"
 
 	"golang.org/x/text/language"
 	"golang.org/x/text/message"
@@ -113,8 +112,7 @@ type pageCommon struct {
 	walletTabs      *decredmaterial.Tabs
 	accountTabs     *decredmaterial.Tabs
 	keyEvents       chan *key.Event
-	toasts          *[]*toast
-	toastList       layout.List
+	toast           **toast
 	states          *states
 	modal           *decredmaterial.Modal
 	modalReceiver   chan *modalLoad
@@ -300,11 +298,9 @@ func (win *Window) addPages(decredIcons map[string]image.Image) {
 		changePage:              win.changePage,
 		setReturnPage:           win.setReturnPage,
 		refreshWindow:           win.refresh,
-		toastList:               layout.List{Axis: layout.Vertical},
+		toast:                   &win.toast,
 	}
 
-	toasts := make([]*toast, 0)
-	common.toasts = &toasts
 	common.fetchExchangeValue(&common.dcrUsdtBittrex)
 
 	common.wallAcctSelector = &walletAccountSelector{
@@ -400,10 +396,10 @@ func (page pageCommon) refreshPage() {
 }
 
 func (page pageCommon) notify(text string, success bool) {
-	*page.toasts = append(*page.toasts, &toast{
+	*page.toast = &toast{
 		text:    text,
 		success: success,
-	})
+	}
 }
 
 func (page pageCommon) closeModal() {
@@ -417,7 +413,7 @@ func (page pageCommon) closeModal() {
 }
 
 func (page pageCommon) Layout(gtx layout.Context, body layout.Widget) layout.Dimensions {
-	page.handleNavEvents()
+	page.handler()
 
 	return layout.Stack{}.Layout(gtx,
 		layout.Expanded(func(gtx C) D {
@@ -476,25 +472,13 @@ func (page pageCommon) Layout(gtx layout.Context, body layout.Widget) layout.Dim
 		}),
 		layout.Stacked(func(gtx C) D {
 			// global toasts. Stack toast on all pages and contents
-			if len(*page.toasts) == 0 {
+			if *page.toast == nil {
 				return layout.Dimensions{}
 			}
 			gtx.Constraints.Min.X = gtx.Constraints.Max.X
 			return layout.Center.Layout(gtx, func(gtx C) D {
-				return page.toastList.Layout(gtx, len(*page.toasts), func(gtx C, index int) D {
-					t := func(n *toast) layout.Dimensions {
-						inset := layout.Inset{Top: values.MarginPadding20}
-						if index == 0 {
-							inset.Top = values.MarginPadding65
-						}
-						return inset.Layout(gtx, func(gtx C) D {
-							(*page.toasts)[index].Timer(time.Second*3, func() {
-								*page.toasts = (*page.toasts)[1:]
-							})
-							return displayToast(page.theme, gtx, n)
-						})
-					}
-					return t((*page.toasts)[index])
+				return layout.Inset{Top: values.MarginPadding65}.Layout(gtx, func(gtx C) D {
+					return displayToast(page.theme, gtx, *page.toast)
 				})
 			})
 		}),
